requirements.txt
tensorflow==2.16.1
pandas
numpy
streamlit
opencv-python-headless
pillow